+++ R/client_id.R
@@ -53,7 +53,7 @@ ga_clientid_hash <- function(webPropertyId, clientId){
 #' 
 #' # make sure you are authenticated with user deletion scopes
 #' options(googleAuthR.scopes.selected = "https://www.googleapis.com/auth/analytics.user.deletion")
-#' ga_auth(new_user = TRUE)
+#' ga_auth()
 #' 
 #' # a vector of ids
 #' ids <- c("1489547420.1526330722", "1138076389.1526568883")