+++ R/ga_v3_get.R
@@ -12,12 +12,9 @@
 #' @param segment How to segment.
 #' @param samplingLevel Choose "WALK" to mitigate against sampling.
 #' @param max_results Default 100. If greater than 10,000 then will batch GA calls.
-#' @param multi_account_batching If TRUE then multiple id's are fetched together.  
-#'     Not compatible with \code{samplingLevel="WALK"} or \code{max_results>10000}
 #' @param type ga = Google Analytics v3; mcf = Multi-Channel Funels.
 #'
 #' @return For one id a data.frame of data, with meta-data in attributes.  
-#'   For multiple id's, a list of dataframes.
 #' 
 #' @seealso https://developers.google.com/analytics/devguides/reporting/core/v3/
 #' 
@@ -53,12 +50,6 @@
 #'                            metrics = c("sessions", "bounceRate"),
 #'                            dimensions = c("source", "medium"))
 #'                        
-#'  ## multi accounts, pass character vector of viewIds
-#'  ## outputs a list of data.frames, named after the viewId
-#'  multi_gadata <- google_analytics_3(id = c("123456","9876545","765432"),
-#'                                   start="2015-08-01", end="2015-08-02",
-#'                                   metrics = c("sessions", "bounceRate"),
-#'                                    dimensions = c("source", "medium"))
 #'                                 
 #' ## if more than 10000 rows in results, auto batching
 #' ## example is setting lots of dimensions to try and create big sampled data
@@ -106,34 +97,16 @@ google_analytics_3 <- function(id,
                                segment=NULL,
                                samplingLevel=c("DEFAULT", 
                                                "FASTER",
-                                               "HIGHER_PRECISION", 
-                                               "WALK"),
+                                               "HIGHER_PRECISION"),
                                max_results=100,
-                               multi_account_batching = FALSE,
                                type = c("ga", "mcf")) {
+  
+  .Deprecated("google_analytics", package = "googleAnalyticsR")
   
   samplingLevel <- match.arg(samplingLevel)
   start <- as.character(start)
   end <- as.character(end)
   type <-  match.arg(type)
-  
-  if(multi_account_batching){
-    if(max_results>10000){
-      warning("Can't use multi_account_batching=TRUE when max_results>10000. 
-              Setting back to FALSE")
-      multi_account_batching <- FALSE
-    }
-    if(samplingLevel=="WALK"){
-      warning("Can't use multi_account_batching=TRUE when samplingLevel=='WALK'. 
-              Setting back to FALSE")
-      multi_account_batching <- FALSE 
-    }
-    if(length(id) == 1){
-      warning("multi_account_batching=TRUE but only one id specified. 
-              Ignoring.")
-      multi_account_batching <- FALSE 
-    }
-  }
   
   if(!is.null(filters)) {
     filters <- utils::URLencode(filters, reserved = TRUE)
@@ -144,7 +117,7 @@ google_analytics_3 <- function(id,
   dimensions <- paste(sapply(dimensions, checkPrefix, prefix=type), collapse=",")
   
   
-  ga_pars <- list(ids = id[1],
+  ga_pars <- list(ids = id,
                   "start-date"=start,
                   "end-date"=end,
                   metrics=metrics,
@@ -154,47 +127,22 @@ google_analytics_3 <- function(id,
                   segment=segment,
                   samplingLevel=samplingLevel,
                   "start-index"=1,
-                  "max-results"=max_results,
+                  "max-results"=min(10000, max_results),
                   output="json")
   
   ga_pars <- rmNullObs(ga_pars)
   
-  # all_data <- list()
   ga <- gar_api_generator(paste0("https://www.googleapis.com/analytics/v3/data/",
                                  type),
                           "GET",
                           pars_args = ga_pars,
-                          data_parse_function = parse_google_analytics,
+                          data_parse_function = parse_google_analytics3,
                           checkTrailingSlash = FALSE)
   
-  if(multi_account_batching){
-    myMessage("Fetching all ids at same time (max 10 per API call)", level = 3)
-    
-    # set custom batch endpoint
-    op <- options()
-    options("googleAuthR.batch_endpoint" = "https://www.googleapis.com/batch/analytics/v3")
-    all_data <- googleAuthR::gar_batch_walk(ga,
-                                            walk_vector = id,
-                                            gar_pars = ga_pars,
-                                            pars_walk = 'ids',
-                                            data_frame_output = FALSE,
-                                            batch_size = 10)
-    all_data <- unlist(all_data, recursive = FALSE, use.names = FALSE)
-    options(op) # reset options
-    names(all_data) <- id
-    
-  } else {
-    all_data <- loop_ids(id=id, 
-                         ga_pars=ga_pars, 
-                         samplingLevel=samplingLevel, 
-                         max_results=max_results, 
-                         ga=ga)
-  }
-
-  
-  all_data
-  
+ results <- gar_api_page(ga, 
+                         page_method = "url", 
+                         page_f = function(x) attr(x, "nextLink"))
+ 
+ Reduce(dplyr::bind_rows, results)
+ 
 }
-
-
-